requirements.txt
certifi==2019.9.11
cffi==1.13.2
cryptography==2.8
emoji==0.5.4
future==0.18.2
pycparser==2.19
PySocks==1.7.1
python-telegram-bot==12.2.0
six==1.13.0
tornado==6.0.3
